shared/requirements.txt
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
pyyaml>=6.0,<7.0
pandas>=2.0,<3.0
numpy>=1.24,<2.0
pyarrow>=14.0,<18.0
redis>=5.0,<6.0
